# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: MacQuiz v2 API
  description: Quiz platform for admins, teachers, and students.
  version: 0.8.0
servers:
  - url: /
paths:
...
  /api/v1/attempts/{id}:
    get:
      operationId: resumeAttempt
      summary: Resume an attempt (owner)
      description: >-
        The saved answers, the deadline, and the current server time - the
        client countdown is cosmetic, driven by server deadline plus a
        clock-offset estimate. Owner-only; anyone else reads 404.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The full player payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttemptDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/attempts/{id}/answers/{questionId}:
    put:
      operationId: saveAnswer
      summary: Autosave one answer (owner)
      description: >-
        Idempotent upsert on (attempt_id, question_id); the question must
        belong to the snapshot version this attempt pinned. Refused with 409
        once now() passes deadline_at + 5 s grace or the attempt has left
        in_progress - the same gate that guards submit.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: questionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveAnswerRequest"
      responses:
        "200":
          description: The stored answer plus the deadline and server clock.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveAnswerResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/AttemptConflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /api/v1/attempts/{id}/submit:
    post:
      operationId: submitAttempt
      summary: Submit the attempt (owner)
      description: >-
        The manual leg of the idempotent submit funnel - whatever is
        autosaved becomes the submission. Submitting an already-submitted
        attempt answers 200 with the unchanged terminal state, so a retried
        request or a race against the deadline job resolves cleanly.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The attempt in its terminal state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttemptResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/AttemptConflict"
  /api/v1/attempts/{id}/result:
    get:
      operationId: getAttemptResult
      summary: The released review for an attempt (owner)
      description: >-
        The score, the per-question grading, and - only here, per the
        documented exception to answer-key isolation - the correct answers.
        Refused with 409 RESULTS_NOT_RELEASED until the quiz's results are
        released and this attempt is graded. Owner-only; anyone else reads
        404. Questions come back in the same per-attempt order the player
        showed.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The released result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttemptResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/ResultsNotReleased"
...
components:
  responses:
    Unauthorized:
      description: Missing, invalid, or expired credentials.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationFailed:
      description: Body-level validation failed; see fields.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: The policy denies this action for the caller's role.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: >-
        No such resource. Also the answer for resources the caller is not
        assigned to, so existence is never leaked.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    AttemptConflict:
      description: >-
        The attempt state refuses this action; the client switches on code -
        QUIZ_NOT_LIVE (start outside the window), ATTEMPT_LIMIT_REACHED
        (max_attempts exhausted), ATTEMPT_DEADLINE_PASSED (write after
        deadline + grace), ATTEMPT_ALREADY_SUBMITTED (autosave against a
        terminal attempt), or ATTEMPT_KICKED (show the lockout screen).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ResultsNotReleased:
      description: >-
        RESULTS_NOT_RELEASED - the quiz's results are not released yet (or
        this attempt's grading has not landed); the client shows a "results
        pending" state.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: >-
            Stable machine-readable code from the docs/04-api.md vocabulary,
            e.g. INVALID_CREDENTIALS, UNAUTHENTICATED, VALIDATION_FAILED,
            RATE_LIMITED, PASSWORD_CHANGE_REQUIRED, NOT_FOUND.
        message:
          type: string
        fields:
          type: object
          additionalProperties:
            type: string
          description: Per-field messages for VALIDATION_FAILED.
...
    QuestionBody:
      type: object
      required: [text]
      properties:
        text:
          type: string
        image_ref:
          type: string
    QuestionOption:
      type: object
      required: [key, text]
      properties:
        key:
          type: string
        text:
          type: string
...
    Guardrails:
      type: object
      description: >-
        Per-quiz anti-cheat config, snapshotted with the question set at
        publish so rules cannot change under a student mid-window.
      required:
        [fullscreen, focus_tracking, block_clipboard, max_violations,
         violation_action]
      properties:
        fullscreen:
          type: string
          enum: [off, warn, count]
        focus_tracking:
          type: string
          enum: [off, warn, count]
        block_clipboard:
          type: boolean
        max_violations:
          type: integer
          minimum: 1
          maximum: 100
        violation_action:
          type: string
          enum: [flag, auto_submit, notify]
...
    Attempt:
      type: object
      description: >-
        The student-facing attempt shape. The owner is always the caller and
        the score is withheld until the results-release policy lands with
        grading, so neither appears here.
      required:
        [id, quiz_id, attempt_no, quiz_version, started_at, deadline_at,
         submitted_at, submit_kind, status, violation_count]
      properties:
        id:
          type: string
          format: uuid
        quiz_id:
          type: string
          format: uuid
        attempt_no:
          type: integer
        quiz_version:
          type: integer
          description: Pins the snapshot the student saw.
        started_at:
          type: string
          format: date-time
        deadline_at:
          type: string
          format: date-time
          description: >-
            least(started_at + duration, quiz.ends_at); precomputed once at
            start, the single column every write validates against.
        submitted_at:
          type: string
          format: date-time
          nullable: true
        submit_kind:
          type: string
          enum: [manual, auto, forced, kicked]
          nullable: true
        status:
          type: string
          enum: [in_progress, submitted, graded, kicked]
        violation_count:
          type: integer
    AttemptQuestion:
      type: object
      description: >-
        One snapshot question as the player sees it. The answer key is
        structurally absent; when the quiz shuffles, positions are
        renumbered densely in the per-attempt order.
      required: [id, position, type, body, points]
      properties:
        id:
          type: string
          format: uuid
        position:
          type: integer
        type:
          type: string
          enum: [single, multi, truefalse, short]
        body:
          $ref: "#/components/schemas/QuestionBody"
        options:
          type: array
          items:
            $ref: "#/components/schemas/QuestionOption"
        points:
          type: number
    AttemptAnswer:
      type: object
      required: [question_id, response, time_spent_ms, saved_at]
      properties:
        question_id:
          type: string
          format: uuid
        response:
          description: >-
            The autosaved response; shape depends on question type - the
            option key (single), option keys (multi), a boolean (truefalse),
            or a string (short).
        time_spent_ms:
          type: integer
        saved_at:
          type: string
          format: date-time
    AttemptDetail:
      type: object
      description: >-
        The full player payload - attempt, quiz identity, snapshotted
        guardrails, questions without the answer key, saved answers, and
        the server clock for the cosmetic countdown.
      required: [attempt, quiz_title, guardrails, questions, answers, now]
      properties:
        attempt:
          $ref: "#/components/schemas/Attempt"
        quiz_title:
          type: string
        guardrails:
          $ref: "#/components/schemas/Guardrails"
        questions:
          type: array
          items:
            $ref: "#/components/schemas/AttemptQuestion"
        answers:
          type: array
          items:
            $ref: "#/components/schemas/AttemptAnswer"
        now:
          type: string
          format: date-time
          description: Current server time, for the clock-offset estimate.
    SaveAnswerRequest:
      type: object
      required: [response]
      properties:
        response:
          description: The response value; shape depends on question type.
        time_spent_ms:
          type: integer
          minimum: 0
          description: Accumulated focus time on this question, for analytics.
    SaveAnswerResponse:
      type: object
      required: [answer, deadline_at, now]
      properties:
        answer:
          $ref: "#/components/schemas/AttemptAnswer"
        deadline_at:
          type: string
          format: date-time
        now:
          type: string
          format: date-time
    AttemptResponse:
      type: object
      required: [attempt]
      properties:
        attempt:
          $ref: "#/components/schemas/Attempt"
...
    ResultQuestion:
      type: object
      description: >-
        One snapshot question in the released review - what was asked, what
        the student answered, the key, and what it earned. Unanswered
        questions keep response and is_correct null.
      required:
        [id, position, type, body, correct, points, response, is_correct,
         points_awarded]
      properties:
        id:
          type: string
          format: uuid
        position:
          type: integer
        type:
          type: string
          enum: [single, multi, truefalse, short]
        body:
          $ref: "#/components/schemas/QuestionBody"
        options:
          type: array
          items:
            $ref: "#/components/schemas/QuestionOption"
        correct:
          description: >-
            The answer key - exposed here only, after close and release
            (the documented exception to answer-key isolation).
        points:
          type: number
        response:
          nullable: true
          description: The autosaved response; shape depends on question type.
        is_correct:
          type: boolean
          nullable: true
        points_awarded:
          type: number
    AttemptResult:
      type: object
      description: >-
        The released results payload for one attempt. The attempt keeps its
        usual score-free shape; the released score rides at the top level.
      required:
        [attempt, quiz_title, score, max_score, released_at, questions]
      properties:
        attempt:
          $ref: "#/components/schemas/Attempt"
        quiz_title:
          type: string
        score:
          type: number
        max_score:
          type: number
        released_at:
          type: string
          format: date-time
        questions:
          type: array
          items:
            $ref: "#/components/schemas/ResultQuestion"